refs #1886: support multi-value attributes in a hackish kind of way
the old implementation assumed that 1 (and only 1) value was returned
from a method invocation.  while this is valid in the case of "int
getLoggerLevel()" or "long calcUptime()" it's not valid in the case of
things like "float getCpuStates()" or "Map<String, Integer>
getDatabasePoolStats()"

the new implementation allows you to provide a List<String> of keys
that correspond to the objects returned.  it's a little dangerous
because we're totally blowing away type safety but it works..


git-svn-id: http://dev.zenoss.com/svnint/trunk/core@19601 f5df01ce-0320-0410-9ed9-ddcd13c732cc