Merging messaging branch into the trunk.
This branch adds Growl-like notifications to the UI, and a versatile backend for storing and delivering messages. It also includes a portlet for displaying persistent messages.

Messages may be sent to the browser directly, in which case they do not persist, or to a user (or all users), in which case they are stored in the database.

Code was reviewed by Chad.

git-svn-id: http://dev.zenoss.com/svnint/trunk/core@25124 f5df01ce-0320-0410-9ed9-ddcd13c732cc