Fixes #29858, Fixes #29863: Fix a bug that was introduced in changeset: [48749] where rulebuilder fields did not apply correctly to the config. This was bad rules to be generated which would be impossible to match in event processing (#29858). Also fixes components for several rulebuilder subjects so they use better combo boxes (#29863).
Update the device router to return integers for production states; also update the device IInfo object to not change the type of the productionState property. Change the default production state when adding a device to use an integer now instead of a string.

Update the device router to return integers for device priority; also update the device IInfo object to not change the type of the priority property. Change the default priority when adding a device to use an integer now instead of a string. Make sure combo boxes in the rule builder that use DirectStores are using smarterSetValue so that loading a rule correctly sets the value on the combo boxes on the rules.

Reviewed by jhanson.


git-svn-id: http://dev.zenoss.com/svnint/trunk/core@53077 f5df01ce-0320-0410-9ed9-ddcd13c732cc