Add a basic test to find undefined names in code
There are a lot of potential bugs in Zenoss code, which can be found with
the static code analyser, so there is a sense to add this check to tests
and take care, that this test pass.

Added a simple proof-of-concept test, which shows undefined variables
and wildcard imports in code. Used flake8 [0] as a static code analyser
and tox [1] as a test runner.

At the moment this tool detects a lot of undefined names and wildcard
imports in code [2]. I want to believe, that this will be fixed someday.

[0] - https://flake8.readthedocs.org/
[1] - https://tox.readthedocs.org
[2] - https://dev.zenoss.com/tracint/pastebin/9967