Fix packZProperties/packZProperties_data clashing.
This corrects a problem where the type for a ZenPack's zProperty as
defined in its ZenPack.packZProperties property would be erased if the
same zProperty was also defined in ZenPack.packZProperties_data without
setting the type. This could end up being a common occurrence once
packZProperties_data becomes more widely known.

This fix prevents the erasure and makes packZProperties win any
disagreements in an attempt to provide the best backwards compatibility.

This problem can also easily be worked around in the ZenPack by adding a
type into the packZProperties_data for each zProperty in that dict.

Fixes ZEN-20359.