Aggregate events for all top-level organizers.
This changes the behavior of what events will be shown in the contextual
event consoles for the Devices, Groups, Systems and Locations organizers
on the Infrastructure view.

Prior to this change there was a special case that would cause all
events associated with a device to be shown for the Devices organizer,
but the event consoles for Groups, Systems and Locations would always be
empty.

This changes the behavior so all of the top-level device organizer event
consoles work the same way. They all show the aggregation of all events
for all of their children. This means that the event console for Groups
will show events for all devices that are in one or more group.

Fixes ZEN-19758.