Show correct events when context changes.
There was no code in place to refresh the event console contents when a
different organizer was chosen on the infrastructure page. This resulted
in seeing events for the last organizer you were looking at until the
console refreshed itself, or your manually refreshed it.

This change adds a "beforeshow" listener to the infrastructure
contextual event console that loads the proper contextual events before
the event console is shown.

Fixes ZEN-19758.