Add a static code analyzer check for DataCollector
Static code analyzers can be used to check code style, undefined and
unused names and so on - see [0], so there is a sense to add such check
as a part of developers testing.

Used flake8 [1] as a static code analyser and tox [2] as a test runner.

The list of allowed checks:
- F401  module imported but unused
- F403  ‘from module import *’ used; unable to detect undefined names
- F821  undefined name name
- F823  local variable name ... referenced before assignment

[0] - http://flake8.readthedocs.org/en/latest/warnings.html
[1] - https://flake8.readthedocs.org/
[2] - https://tox.readthedocs.org