Fix permission checks so that they fall through correctly
If role based permissions granted access to a button, the
button was always enabled.  This change allows other states
(like no devices being selected) to disable controls even
if the role has access to the function.  This check-in also
fixes a permission error for the action and commands dropdown
menus so they get disabled properly when permissions change.