Update zenpass to work correctly.
We no longer use zopectl, so we need to send control-c to the
runzope command to restart it.  We also give a help message
if they give arguments (like --help) to prevent passthrough to
the zpasswd.py Zope command, and check to make sure it looks
like they're in a zope shell.