Fix various issues in compare routine
  The json deserialization in serviced does a case-insensitive match
on the field name, choosing the last match encountered for any field.
The keys are sorted in service-migration, so in the case of duplicate
values for a field the "most lower-case" is the last and therefore wins.

  This is a problem because the migration SDK has a bug in which some of
the fields of the LogConfig are written in both lowercase and camelcase.
Since serviced supports this, and the migration SDK uses it, it needs to
be handled here.

  Because the dictionary lookup is modified, the error reporting code
needs to change as well.  (Formerly it would look up each object in the
path by the key; now that key is lowercase and may not be in the
dictionary.)  Instead we return a "Diff" object with actual and expected
values.

  Finally, while modifying the code I took the opportunity to change the
string comparison so that it outputs a unified diff iff the strings are
multi-line; for single line strings we just do a string comparison.
This results in more meaningful output when single line strings differ.

+ Perform case-insensitive key matching when comparing dictionaries
+ Return a Diff object with actual and expected values
+ Return unified diff object only in case of multiline string;single
  line strings treated normally.
+ For dictionaries when keys are not present, uses a sentinel value
  in the Diff return value for the missing key
+ Add unit tests for the various cases