Fix output json for various unit tests
As a result of fixes made to common.compare, some existing migration
unit tests broke.  The problems consisted of cases in which formerly a
Null value was not distinguished from a missing key, and cases where a
default value had been added to the expected output (as a result of a
round trip through serviced) which was not present in the modified json.

This PR modifies the expected output from the migration to match the
actual output.