Handle connection reset during getresponse
  The CC server resets the connection when an unauthenticated POST
request is made.  Depending on when during the request lifecycle the
connection is reset, we can get either an URLError with a socket.error
as the reason, or a naked socket.error.  In either case, the
socket.error.errno indicates that the connection was reset with an
errno of ECONNRESET (104).
  When we get a connection reset exception, assume that the reset was
caused by lack of authentication, login, and retry the request.

Fixes ZEN-24636