Always transform ZenPack version for egg file
When looking for the egg (or zip) of a ZenPack, the version string must
convert its first dash into an underscore to match version string used in
the naming of the egg (or zip) file.  This change moves that transform into
the method that performs the search rather relying the caller to transform
it first.

This should finish fixing ZEN-24265.