ZEN-17882: update MIB module instead of remove.
When we install new MIB file zenmib will just remove the old one
and install new one, despite fact that previous MIB file was newer
or there was some nodes or notification added manually, this will
cause a crash of eventClassMappings.
If that MIB file is in the system, take that object and just update
it.