Fix ZEN-26944 pt2 adds more logging info, fixes race between mainLoop & checkStopadd collector loop counter, more logging, check if mainLoopGotDeviceList (#2154)
* add collector loop counter, more logging, check if mainLoopGotDeviceList

- added a collector loop iteration counter
- add more logging to `fillCollectionSlots()`,
`processClientFinished()` & `checkStop()` to follow progress of modeling
run
- The root cause of the defect seems to be that `checkStop` signals
the end of a collector loop before `mainLoop` can receive and set the
modelers deviceList on which to collect. To remedy, we set a flag to let
`checkStop` know if the `deviceList` was set for the current modeling
run. If this fix doesn't work, the additional logging information will
still provide the user with more info to determine what work is being
done during a given modeling run.

* initialize ZenModeler.mainGotDeviceList in constructor