Prevent duplicate OIDs.
MIB OIDs may be added to Zenoss via ZenPacks, the zenmib utility, and via the web UI.  This change will delete a pre-existing OID from the database before a new instance of the OID is added to prevent duplicates.  The web UI will prompt the user to confirm the replacement action.