ZEN-23132: change debug to error log and raise an exception
When we were adding CC host first as /Server/SSH/Linux or
/Server/Linux device class and then using the VIP to
the /Control Center device class The CC device will not
appear even though the job completes successfully.

A device can't be added since the VIP was already assigned
to device when it was first added and modeled as
/Server/SSH/Linux or /Server/Linux. An error wasn't raised
since exception was catched. Now give an error log and raise
an exception for job to be failed.