ZEN-24255: add new rpn field to threshold
It was confusing for user what value should he enter in min/max
field, since after that value was calculated with the same rpn
of the datapoint threshold is related to.
Added new field to to Threshold UI which show the rpn expression
that will be used to threshold min/max value. This will help to
adopt the value to rpn expression.