ZEN-27051: add new button to show threshold with rpn
It was confusing for user what value should he enter in min/max
field, since after that value was calculated with the same rpn
of the datapoint threshold is related to.
Added new button to to Threshold UI which will show table with
the values of the threshold after rpn was used.