Allow upgrading of auth0 installation
Previously the auth0 plugin installation would abort if it detected that
it was already installed.  After this change, if the version of the
installed plugin does not match the current version, it will delete the
existing version and reinstall.  This allows the code which registers
interfaces and reorders the plugins to run, even if a previous version
of the plugin exists.