ZEN-29466: formatting version from egg and ZODB before restore
After upgrade, some of the dev zenpacks was left in a broken state.
The problem was in version naming of the zenpack egg and version
naming from ZODB. At the time of upgrade, those two versions were
formatted differently and Zenpack was not able to restore, so it was
skipped and left in a broken state.