Store only the relevant fields of the id token
Previously, the entire id token was placed in session storage.  Each
request would JWT decode the token to extract the userid and check for
expiration.

After this change, the relevant fields of the token are extracted when
the token is received and those fields are placed in session storage.
The token expiration is explicitly checked, rather than implicitly as
part of the JWT token decode.  All of the relevant fields are placed
into a single object in the session for easier bookkeeping.