Ensure that AccountLocker is first PAS-plugin
Both AccountLocker and Auth0 include code to movethemselves to the front
of the PAS plugin list.  We actually want AccountLocker to be first, so
that it can lock users out before Auth0 can log them in.

This change ensures that Auth0 is added before AccountLocker so that
AccountLocker will be moved to the front of the list.  Additionally, we
delete existing AccountLocker objects if they are old in order to ensure
that that "move to top" function will be called; previously, if the
plugin already existed this routine would be skipped.