Implement IRolesPlugin for Auth0
Previously, we had hacked in an event handler to set the roles for any
newly created user.  However, this approach stopped working when we
removed the AuthoLogin page.

The new approach is to implement the PAS plugin IRolesPlugin interface.
This allows us to specify the roles for users as they are created.  In
the long run the intent is to extract the roles from the Auth0 token
and store them in the session data.  Then, when it comes time to assign
the roles, we can extract them from the session data (via the provided
request).

Also removed the obsolete event handler.