Check the token expiration cookie, if set
If a user logs into ZING, then opens CZ, CZ will cache the access token.  If
the user closes the CZ window and logs in to ZING with a different user, then
opens the CZ window again, CZ uses the cached session data.  We need to check
to make sure the cached data is still valid, without parsing the JWT with
each request -- if the 'accessTokenExpiration' cookie is set, use this to make
sure our session expiration matches.  If it doesn't, then the token has changed
and we need to parse/store the token again.