ZEN-30469: The test regex window is not fetching process list
When we have the remote collector getProcessList should be
run on that collector, but it is not.
Logical operator that is used to set the modeler collector
is not right. We should use remote modeling in the case
when the collector is not equal to localhost and do not care
about the path to zenmodelerPath.

(cherry picked from commit 1c7568fbb84775d476d0a0a854fb916ec3854446)