RRDDataPoint: add store, forward, tags properties
Add support for all datapoints to have three new capabilities related to
metric tagging.

- store
- forward
- tags

The store property will result in the {"no-store": "true"} tag being
added to published metrics for the datapoint when set to False, and no
additional tags set when set to True. True is the default to retain
backwards compatibility for all existing datapoints, and all new
datapoints by default.

The forward property will result in the {"no-forward": "true"} tag being
added to published metrics for the datapoint when set to False, and no
additional tags set when set to True. True is the default to retain
backwards compatibility for all existing datapoints, and all new
datapoints by default.

The tags property allows arbitrary additional tags to be published for
the datapoint's metrics. Multiple tags can be specified in the following
format (one per line):

    name1:value2
    name2:value2

These tags will be evaluated as TALES expressions so examples like the
following will also work.

    deviceId:${device/id}
    hereId:${here/id}
    interfaceName:${here/interfaceName}
    totalMemory:${here/hw/totalMemory}

These changes only immediately add this support for metrics collected by
"platform" (non-ZenPack) collector services. Specifically the following
collectors.

- zencommand
- zenperfsnmp
- zenping
- zenprocess

All collector services that are part of ZenPacks must be updated to use
this new capability individually. Specifically by updating their
HubService to pass the per-datapoint tags down to the collector service,
and their collector services' calls to writeMetricWithMetadata to
include those tags as the extraTags parameter. See the
CommandPerformanceConfig.py and zencommand.py changes as a reference for
the kind of updates that are required.

Implements ZEN-30906.