ZEN-30743: allow dot in the datasource and datapoint name
A while ago rules for datasource name was added, that rule
was disabling a form if we have any special character in the
name except underscores and dashes since datasource name is
part of the metric name in OpenTSDB that was breaking graphs.
According to OpenTSDB documentation, dots are also acceptable in
metric names.