ZEN-30592: remove buildTextLabel functionality
When a user uses page size filter on the table view and sorts
the column that has int value in it will fall with a traceback.
Also, values from the sorted column are used as a page name on
pagination select element, which is not logically right, there
should be only the page numbers 1,2,3 ....
Remove that functionality since it is confusing and not logical.