ZEN-31177: send to UI asterisks instead of valid password
User password for SMTP server is accessible in browser dev tool
by just selecting an input field, the input field value attribute
will show the password as clear text.
Set the value attribute of the input field to masked password under
asterisks. When the form on the UI is submited save the password
only if its new. When SMTP password field is in focus clear it
to prevent updating old password since it can cause wrong password
situation(****  will be saved like ****56)