Allow Twisted to unserialize unregistered classes.
A custom SecurityOptions implementation is now given to the Perspective Broker.  The new implementation is more permissive regarding what types and classes are allowed for serialization.  The only restriction is that classes must implement the IUnjellyable interface.  This is a defacto requirement as Twisted is written to assume that it has been implemented.  In practice this will not be a concern because any class inheriting the RemoteCopy class implements the IUnjellyable interface.

Fixes ZEN-31574.