ZEN-32689 Update 'mariadb_answering' healthcheck in zeneventserver
This commit replaces the name of healthcheck 'mariadb_answering' with 'mariadb_events_anwering' for clarity.
Also, the script for healthcheck was changed, it used the credentials for zodb when zeneventserver actually connects to the zep.
These changes fix the problem when healthcheck is failed after changing a password for 'zenoss' user in mariadb-events