Set job to FAILURE if ERROR in its log file.
Set a job to the FAILURE state if an ERROR log message is found in the job's log file.  Many jobs do not propagate their errors into zenjobs (celery), so zenjobs has to examine their log file to determine the jobs' final state.

Also fixed an issue with the Redis wrapper (storage) when updating a job record with nothing but None valued fields.

Fixes ZEN-33036.