Fix adding new user to group without permission
This way of adding a previously non-existent user to a group can result
in breaking things in a CZ. Only users with the CZAdmin role, and
therefore the "Manage Users" permission should be allowed to add new
users to a group. Users with the ZenManager role will still be allowed
to add existing users to groups.

Fixes ZING-14314.