Add support for multiple instances of collector daemons, using a rendezvous hashing algorithm
The traditional hashing we did in RM was based on a sequential instance number, which
wasn't really practical in the more dynamic scenario we had in mind with collectors being
scaled up and down in k8s.

Now the hashing is done based on container ID, and the rendezvous algorithm gives consistent
results as long as the same list of IDs is used.

This requires us to keep track of collector listener connections as they come and go,
and making sure that a consistent collector list is used by any process that needs
to perform the hash.

In addition we take extra care to push config changes whenever the responsible collector
changes for a given device, as will happen as they are added and removed.