Fix UI tracebacks messages for event transforms (#4127)
* Fix UI tracebacks messages for event transforms

Fixes ZEN-32905.

To fix type errors related to getSecurityPermissions() and
getTransformTree() which were observable for the customer through the
UI, the default value for event class was set. After changes when the
customer opens Events -> Even Classes and chose Transforms in the
drop-down menu before selecting an event class, no errors will occur,
instead of them the transforms for Events will be shown.

* Fix UI issues for event transforms

Fixes ZEN-32905.

To improve the solution of type errors related to
getSecurityPermissions() and getTransformTree() which were observable
for the customer through the UI, the default value for event class was
set on the front-end side instead of the back-end side.

* Fix getSecurityPermissions() issue for UI

Fixes ZEN-32905.

To improve the solution of type error related to
getSecurityPermissions() which was observable for the customer through
the UI, the default value for event class was set on the front-end side
instead of the back-end side.