Handle and log the UnicodeDecodeError
Fixes ZEN-32208.

To handle the issue with UnicodeDecodeError the decoding procedure was
modified by adding additional checks. The procedure tries to decode
values using zCollectorDecoding if it is defined, otherwise, it tries to
apply the utf-8 and then latin-1encodings simultaneously logging
information about this. If decoding of the value in strict mode causes
errors, then it will be done in ignore mode, therefore non-Unicode
symbols will be ignored.