Fix hasattr_path within sendEvent for SshClient
Fixes ZEN-31888.

To fix the issue the logic of nested function hasattr_path() within
sendEvent() for SshClient was changed and simplified. Since the
object_root objects, which is an argument for the hasattr_path()
function, has a defined structure that contains the attribute 'factory'
and doesn't contain any other parts of splitting using "." for
"factory.hostname", "conn.factory.hostname",
"factory.datacollector.sendEvent", "factory.sendEvent",
"datacollector.sendEvent", "conn.factory.datacollector.sendEvent", which
are used as a value for the parameter "path". Consequently the logic of
hasattr_path() was changed to inclusion check of "factory" to
path.split(".").