Feature/zen 33996: Fix adding events with an empty summary (#4145)
* Fix TypeError in CancelledError handling in Executor

Fixes ZEN-33996.

The TypeError in CancelledError handling in Executor occurred because
the wrong log string formatting in the CancelledError handling section
in Executop.py was made. The issue was resolved by fixing the string
formatting.

* Fix adding events with an empty summary

Fixes ZEN-33996.

The issue occurred due to improper processing of errors
twisted.internet.defer.CancelledError and
twisted.internet.error.ConnectionLost, which caused adding events with
an empty summary. To fix the problem, additional except-sections for
these errors were added, which handle these errors and send events with
a clear non-empty summary for twisted.internet.error.ConnectionLost.

* Fix the message for logging CancelledError

Fixes ZEN-33996.

To improve the logging of CancelledError the logging mode was changed to
debug and the message was modified by adding an explanation of why the
deferred was canceled.

* Fix the message for error logging

Fixes ZEN-33996.

To improve the logging of errors, the log message generation was
improved.

Co-authored-by: Artem <azamrii@zenoss.com>