Fix permission verification for event management (#4165)
* Fix permission verification for event management

Fixes ZEN-33568.

The issue occurred because the verification of permissions for event
management within the organizers such as Systems, Groups, and Locations,
was performed using the names of general organizers. It didn't take into
account that different kinds of organizers can have the same names and
can have sub organizers. To fix the issue, the verification of event
management permission was changed by adding verification of sub
organizers. In addition, to identify permissions for different
administered objects the usage of the organizer name was changed on the
bread crumb URL path, which makes all identifiers unique.

* Improve the creation of the manage_events_for list

Fixes ZEN-33568.

The creation of the manage_events_for list was improved by the
substitution of the list() to [].