Fix user group permissions management (#4166)
Fixes ZEN-34035, ZEN-34065.

The issue occurred because object names were used for the identification
of administered objects. When two different administered objects, which
are related to different organizers (e. g. Groups and Locations), have
the same name they were unrecognized as different objects. To fix the
removing of administered objects as well as updating roles for
previously added objects, the identification of objects was changed by
using the primary dmd id.