Fix the downloading MIBs to nested organizers
Fixes ZEN-34139.

The issue occurred because the mib organizer path was computed in the
wrong way, i.e. os.path.join("/zport/dmd/Mibs", "/Nested Org") returned
"/Nested Org" because of / at the beginning. To resolve the issue the
way of computing of mib organizer path was changed to string
concatenation.