Fix the creation events with an empty summary (#4211)
Fixes ZEN-28577.

The TypeError in CancelledError handling in Executor occurred because
the wrong log string formatting in the CancelledError handling section
in Executop.py was made. The issue was resolved by fixing the string
formatting. The issue occurred due to improper processing of errors
twisted.internet.defer.CancelledError and
twisted.internet.error.ConnectionLost, which caused adding events with
an empty summary. To fix the problem, additional except-sections for
these errors were added, which handle these errors and send events with
a clear non-empty summary for twisted.internet.error.ConnectionLost.