zendisc used ip assigned to netapp interface
Fixes ZEN-30583.

zendisc creating devices for ip addresses
assigned to a netapp components. There is because of
relation ipaddress <-> manageDevice.
So when we are checking for the device existence
we check manageDevice for that ipaddress.
Since we no longer check device existence
from ipaddress<->interface relation when
creating a device that is why zendisc is discovering
devices with ip that are already assigned to
the netapp components. Additional check was added
to remove such IPs from the discovering list.