The show command displays more device config content.
The show command now displays nearly all data in a device configuration
and its output adjusts to the width of the terminal.

Also, wildcard matching is now used so the commands can be abbreviated
to reduce the amount of required typing.  The wildcard feature has been
extended to list and expire commands.

ZEN-34721