Add `dmd_read_only` flag to `DMD` mixin class.
Add the `dmd_read_only` attribute to tasks using the `DMD` mixin class.
When set to True, the transaction is aborted when the task completes.
By default, `dmd_read_only` is False.

Removed Device.getRRDTemplates doctest because it was dependent on zodb
being in a specific state to succeed.

ZEN-34849